+++ FE/admin_pages/src/components/common/Category.jsx
@@ -56,7 +56,7 @@ const StyledLink = styled(Link)`
 `;
 
 const categories = [
-  { name: "회원관리", link: "/admin/users?page=1" },
+  { name: "회원관리", link: "/admin/users/register" },
   { name: "회원등록", link: "/admin/users" },
   { name: "도서관리", link: "/admin/books?page=1" },
   { name: "도서승인", link: "/admin/books/approve" }
